feat: 무작위 초밥 레일 제공 api 생성 [S12P11B109-94] - 알고리즘에 따른 초밥 제공 - querydsl 설정 - 더미 데이터 생성

+++ backend/sushi/src/main/java/com/ssafy/sushi/domain/user/UserController.java
@@ -19,7 +19,7 @@ public class UserController {
 
     @GetMapping("/user/me")
     public ResponseEntity<ApiResponse<UserInfoResponse>> createUser(@AuthenticationPrincipal UserPrincipal userPrincipal) {
-        Long userId = AuthenticationUtil.getCurrentUserId(userPrincipal);
+        Integer userId = AuthenticationUtil.getCurrentUserId(userPrincipal);
 
         return ApiResponse.success(userService.getUserInfo(userId));
     }
